fix: revamp newsletter ui
