+++ .talismanrc
@@ -1,6 +1,8 @@
 fileignoreconfig:
-- filename: src/app/(private)/(dashboard)/admin/signaux-faible-beta/AdminProductClientPage.tsx
-  checksum: a8924d7f6456185db63a9b4f7c0c013d65eacecd5e696472adf6d1f979b38f83
+- filename: src/components/NewsletterPage/NewsletterPage.tsx
+  checksum: 6b0b2109d112d5302ef35256a2febf53feadf65f5538acee4dce09f0d2da75eb
+version: ""
+96472adf6d1f979b38f83
 - filename: src/app/(private)/(dashboard)/admin/signaux-faibles-beta/AdminProductClientPage.tsx
   checksum: 9c740513497d70e871e4abcad3a9e509ef5b1a0da9db2d0d2d336c19854303b4
 - filename: src/app/(private)/(dashboard)/incubators/[id]/info-form/page.tsx
